+++ researchlib/metrics/classification.py
@@ -50,12 +50,12 @@ class Acc(Matrix):
             # mixup
             y_pred, y_true, y_true_res, lam = loss_input[0], loss_input[1], loss_input[2], loss_input[3]
             _, predicted = torch.max(y_pred, 1)
-            self.correct += (lam * predicted.eq(y_true).sum().float()
-                        + (1 - lam) * predicted.eq(y_true_res).sum().float())
+            self.correct += (lam * predicted.eq(y_true.long()).sum().float()
+                        + (1 - lam) * predicted.eq(y_true_res.long()).sum().float())
         else:
             y_pred, y_true = loss_input[0], loss_input[1]
             _, predicted = torch.max(y_pred, 1)
-            self.correct += predicted.eq(y_true).sum().float()
+            self.correct += predicted.eq(y_true.long()).sum().float()
         self.total += predicted.view(-1).size(0)
         
     def output(self):
